// fe/src/lib/api/generated/model/significantLearningValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type SignificantLearningValue = typeof SignificantLearningValue[keyof typeof SignificantLearningValue];


export const SignificantLearningValue = {
  INDEPENDENCE: 'INDEPENDENCE',
  INTERPERSONAL_RELATIONS: 'INTERPERSONAL_RELATIONS',
  INTELLECTUAL_GROWTH: 'INTELLECTUAL_GROWTH',
  SPIRITUAL_GROWTH: 'SPIRITUAL_GROWTH',
  RESPONSIBILITY: 'RESPONSIBILITY',
  WORKING_UNDER_PRESSURE: 'WORKING_UNDER_PRESSURE',
  TIME_MANAGEMENT: 'TIME_MANAGEMENT',
  SETTING_PRIORITIES: 'SETTING_PRIORITIES',
  OTHER: 'OTHER',
} as const;